+++ examples/functions-airtable-form/gatsby-config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  flags: {
+    FUNCTIONS: true,
+  },
+  siteMetadata: {
+    title: "Airtable Form",
+  },
+  plugins: ["gatsby-plugin-gatsby-cloud"],
+}