+++ packages/gatsby/src/utils/worker/child/index.ts
@@ -1,0 +1,3 @@
+// Note: this doesn't check for conflicts between module exports
+export { renderHTMLProd, renderHTMLDev } from "./render-html"
+export { loadConfigAndPlugins } from "./load-config-and-plugins"